FastAPI2/requirements.txt
requests
pydantic
fastapi
python-dotenv
google-search-results
langchain-community
langchain-pinecone
langchain
pinecone-client
openai
langchain-openai